fix: use node 16 for publish action
Publish currently returns an error. Seems to be because the default
node.js version has been updated in ubuntu@latest to new LTS (18)
which is not building with CURRENT dependencies:

https://stackoverflow.com/questions/69692842/error-message-error0308010cdigital-envelope-routinesunsupported

Seeing as the latest node version that is supported by the GHA runner
is 16 anyway, we might as well use that.

Will have a go at updating the dependencies in a separate change and
see if it works with 18 for ncc and 16 for the published artifact.